separated the Product Card component in the Product Lists Container into it's own component inside the components folder. Tested and working as before when it was inside the ProductLists container.
