+++ src/components/ProductCard/ProductCard.module.scss
@@ -1,0 +1,9 @@
+.ProductCard {
+    // border: 2px solid #222;
+    // padding: 20px;
+    // margin: 10px auto;
+    // background-color: #efefef;
+
+    // height: 100%;
+    
+}
