+++ src/components/ProductCard/index.js
@@ -1,0 +1,2 @@
+import ProductCard from "./ProductCard";
+export default ProductCard;
