+++ src/containers/ProductList/ProductList.module.scss
@@ -1,13 +1,3 @@
 .ProductList {
     
 }
-
-.ProductCard {
-    // border: 2px solid #222;
-    // padding: 20px;
-    // margin: 10px auto;
-    // background-color: #efefef;
-
-    // height: 100%;
-    
-}
